Fix the rest of Pidgin's warnings about -Wsign-compare; remove -Wno-sign-compare from cflags

+++ pidgin/gtkthemes.c
@@ -295,7 +295,7 @@ void pidgin_themes_load_smiley_theme(const char *file, gboolean load)
 			}
 			while  (*i) {
 				char l[64];
-				int li = 0;
+				gsize li = 0;
 				while (*i && !isspace(*i) && li < sizeof(l) - 1) {
 					if (*i == '\\' && *(i+1) != '\0')
 						i++;
